+++ lib/openapi_first/test.rb
@@ -5,7 +5,7 @@ require_relative 'registry'
 
 module OpenapiFirst
   # Test integration
-  module Test
+  module Test # rubocop:disable Metrics/ModuleLength
     autoload :Coverage, 'openapi_first/test/coverage'
     autoload :Methods, 'openapi_first/test/methods'
     autoload :Observe, 'openapi_first/test/observe'
@@ -13,6 +13,7 @@ module OpenapiFirst
     extend Registry
 
     class CoverageError < Error; end
+    class UnknownQueryParameterError < Error; end
 
     # Inject request/response validation in a rack app class
     def self.observe(app, api: :default)
@@ -31,6 +32,11 @@ module OpenapiFirst
 
     def self.configuration
       @configuration ||= Configuration.new
+    end
+
+    def self.registered?(oad)
+      key = oad.key
+      definitions.any? { |(_name, registered)| registered.key == key }
     end
 
     # Sets up OpenAPI test coverage and OAD registration.
@@ -99,34 +105,22 @@ module OpenapiFirst
 
       OpenapiFirst.configure do |config|
         @after_request_validation = config.after_request_validation do |validated_request, oad|
+          next unless registered?(oad)
           raise validated_request.error.exception if raise_request_error?(validated_request)
+
+          check_unknown_query_parameters(validated_request)
 
           Coverage.track_request(validated_request, oad)
         end
 
         @after_response_validation = config.after_response_validation do |validated_response, rack_request, oad|
+          next unless registered?
           raise validated_response.error.exception if raise_response_error?(validated_response, rack_request)
 
           Coverage.track_response(validated_response, rack_request, oad)
         end
       end
       @installed = true
-    end
-
-    def self.raise_request_error?(validated_request)
-      return false if validated_request.valid?
-      return false unless configuration.raise_error_for_request.call(validated_request)
-      return false if validated_request.known?
-
-      !configuration.ignore_unknown_requests
-    end
-
-    def self.raise_response_error?(validated_response, rack_request)
-      return false if validated_response.valid?
-      return false unless configuration.response_raise_error
-      return false unless configuration.raise_error_for_response.call(validated_response, rack_request)
-
-      !configuration.ignore_response?(validated_response)
     end
 
     def self.uninstall
@@ -137,6 +131,44 @@ module OpenapiFirst
       @configuration = nil
       @installed = nil
       @exit_handler = nil
+    end
+
+    class << self
+      private
+
+      def check_unknown_query_parameters(validated_request)
+        return unless validated_request.known?
+
+        unknown_parameters = validated_request.unknown_query_parameters
+        return unless unknown_parameters
+
+        message = unknown_parameters_message(unknown_parameters, validated_request)
+        raise UnknownQueryParameterError, message
+      end
+
+      def unknown_parameters_message(unknown_parameters, validated_request)
+        s = 's' if many?(unknown_parameters)
+        list = unknown_parameters.keys.map(&:inspect).join(', ')
+        "Unknown query parameter#{s} #{list} for #{validated_request.fullpath}"
+      end
+
+      def raise_request_error?(validated_request)
+        return false if validated_request.valid?
+        return false unless configuration.raise_error_for_request.call(validated_request)
+        return false if validated_request.known?
+
+        !configuration.ignore_unknown_requests
+      end
+
+      def many?(array) = array.length > 1
+
+      def raise_response_error?(validated_response, rack_request)
+        return false if validated_response.valid?
+        return false unless configuration.response_raise_error
+        return false unless configuration.raise_error_for_response.call(validated_response, rack_request)
+
+        !configuration.ignore_response?(validated_response)
+      end
     end
   end
 end