+++ spec/test_spec.rb
@@ -490,6 +490,22 @@ RSpec.describe OpenapiFirst::Test do
     end
   end
 
+  describe '.registered?' do
+    it 'returns false for an unknown OAD' do
+      expect(described_class.registered?(definition)).to be false
+    end
+
+    it 'returns true for a globally registered OAD' do
+      OpenapiFirst.register(definition)
+      expect(described_class.registered?(definition)).to be true
+    end
+
+    it 'returns true for an OAD registered in Test' do
+      OpenapiFirst::Test.register(definition)
+      expect(described_class.registered?(definition)).to be true
+    end
+  end
+
   describe 'handling invalid requests' do
     let(:definition) do
       OpenapiFirst.parse(
@@ -539,6 +555,19 @@ RSpec.describe OpenapiFirst::Test do
       app.call(Rack::MockRequest.env_for('/stuff/nostring'))
 
       expect(described_class::Coverage.result.coverage).to eq 50
+    end
+
+    it 'tracks the request' do
+      expect(described_class::Coverage).to receive(:track_request)
+
+      app.call(Rack::MockRequest.env_for('/stuff/nostring'))
+    end
+
+    it 'does not track the request for for unregistered OADs' do
+      expect(described_class).to receive(:registered?).twice.and_return(false) # once for request, once for response tracking
+      expect(described_class::Coverage).not_to receive(:track_request)
+
+      app.call(Rack::MockRequest.env_for('/stuff/1'))
     end
   end
 
@@ -603,6 +632,64 @@ RSpec.describe OpenapiFirst::Test do
           app.call(Rack::MockRequest.env_for('/unknown'))
         end.not_to raise_error
       end
+    end
+  end
+
+  describe 'handling unknown query parameters' do
+    let(:definition) do
+      OpenapiFirst.parse(
+        {
+          'openapi' => '3.1.0',
+          'paths' => {
+            '/stuff' => {
+              'get' => {
+                'parameters' => [
+                  {
+                    'name' => 'color',
+                    'in' => 'query',
+                    'required' => true,
+                    'schema' => {
+                      'enum' => %w[red green]
+                    }
+                  }
+                ],
+                'responses' => {
+                  '200' => {
+                    'descrition' => 'Ok'
+                  }
+                }
+              }
+            }
+          }
+        },
+        filepath: 'somefile'
+      )
+    end
+
+    let(:app) do
+      described_class.app(
+        ->(_env) { [200, { 'content-type' => 'application/json' }, ['foo']] },
+        spec: definition
+      )
+    end
+
+    before(:each) do
+      described_class.setup do |test|
+        test.register(definition)
+        test.report_coverage = false
+      end
+    end
+
+    it 'raises an error' do
+      expect do
+        app.call(Rack::MockRequest.env_for('/stuff?color=red&unknown=12'))
+      end.to raise_error OpenapiFirst::Test::UnknownQueryParameterError, 'Unknown query parameter "unknown" for /stuff?color=red&unknown=12'
+    end
+
+    it 'raises another error if request is unknown' do
+      expect do
+        app.call(Rack::MockRequest.env_for('/unknown?unknown=12'))
+      end.to raise_error OpenapiFirst::NotFoundError
     end
   end
 
@@ -756,6 +843,54 @@ RSpec.describe OpenapiFirst::Test do
           app.call(Rack::MockRequest.env_for('/roll', method: 'POST'))
         end.not_to raise_error
       end
+    end
+  end
+
+  describe 'handling valid responses' do
+    let(:definition) do
+      OpenapiFirst.parse(YAML.load(%(
+        openapi: 3.1.0
+        info:
+          title: Dice
+          version: 1
+        paths:
+          "/roll":
+            post:
+              responses:
+                '200':
+                  content:
+                    application/json:
+                      schema:
+                        type: integer
+                        min: 1
+                        max:
+      )))
+    end
+
+    let(:app) do
+      described_class.app(
+        ->(_env) { [200, { 'content-type' => 'application/json' }, ['1']] },
+        spec: definition
+      )
+    end
+
+    before(:each) do
+      described_class.setup do |test|
+        test.register(definition)
+      end
+    end
+
+    it 'tracks the response' do
+      expect(described_class::Coverage).to receive(:track_response)
+
+      app.call(Rack::MockRequest.env_for('/roll', method: 'POST'))
+    end
+
+    it 'does not track the response for for unregistered OADs' do
+      expect(described_class).to receive(:registered?).twice.and_return(false) # once for request, once for response tracking
+      expect(described_class::Coverage).not_to receive(:track_response)
+
+      app.call(Rack::MockRequest.env_for('/roll', method: 'POST'))
     end
   end
 end