+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## Unreleased
 
+## 3.1.1
+
+- Changed: Return uniqe errors in default error responses
+
+## 3.1.0
+
+### openapi_first/test
+
+#### Changed
+- OpenapiFirst::Test now raises `OpenapiFirst::Test::UnknownQueryParameterError` when it sees unknown query parameters. Note that `OpenapiFirst` ("core") still allows unknown query parameters.
+- OpenapiFirst::Test does not track requests/responses unless the OAD was registered via `OpenapiFirst::Test.register` (or `OpenapiFirst.register`)
+
 ## 3.0.1
 - Add missing gem dependency "drb", which is no longer installed by default with newer rubies. This is used in openapi_first/test to make parallel tests work.